添加架包工具类，实现新品首页的下载及下拉刷新，上拉加载。

+++ app/src/main/java/cn/ucai/fulicenter/bean/NewGoodsBean.java
@@ -44,7 +44,7 @@ public class NewGoodsBean {
     private String colorName;
     private String colorCode;
     private String colorUrl;
-    private int addTime;
+    private long addTime;
     private boolean promote;
     private PropertiesBean propertiesBean;
 
@@ -184,11 +184,11 @@ public class NewGoodsBean {
         this.colorUrl = colorUrl;
     }
 
-    public int getAddTime() {
+    public long getAddTime() {
         return addTime;
     }
 
-    public void setAddTime(int addTime) {
+    public void setAddTime(long addTime) {
         this.addTime = addTime;
     }
